# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API learning_platform",
        "title": "Learning Platform API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/courses/{course_id}/enroll": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enrollments"
                ],
                "summary": "Đăng ký khóa học",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID khoá học",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Đăng ký học thành công",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Khóa học đã đầy hoặc đã đăng ký trước đó",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Chưa đăng nhập hoặc token không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enrollments"
                ],
                "summary": "Hủy đăng ký học",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID khóa học",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hủy đăng ký thành công",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Không tìm thấy đăng ký học",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Chưa đăng nhập hoặc token không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/courses/{course_id}/lessons": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lessons"
                ],
                "summary": "Danh sách bài học của khóa học",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID khóa học",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Số trang",
                        "name": "page_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Số lượng mỗi trang",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách bài học",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/models.Lesson"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ID khóa học không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Chưa đăng nhập hoặc token không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lessons"
                ],
                "summary": "Tạo bài học mới",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID khóa học",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Thông tin bài học",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.lessonRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tạo bài học thành công",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dữ liệu đầu vào không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Chưa đăng nhập hoặc token không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/enrollments/{enrollment_id}/lessons/{lesson_id}/progress": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Progress"
                ],
                "summary": "Cập nhật tiến độ khóa học",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID đăng ký khóa học",
                        "name": "enrollment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID bài học",
                        "name": "lesson_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cập nhật tiến độ thành công",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ID đăng ký khóa học hoặc ID bài học không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Chưa đăng nhập hoặc token không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/enrollments/{enrollment_id}/progress": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Progress"
                ],
                "summary": "Xem tiến độ học tập",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID đăng ký khóa học",
                        "name": "enrollment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tiến độ học tập",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "progress": {
                                    "$ref": "#/definitions/models.CourseProgress"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ID đăng ký khóa học không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Chưa đăng nhập hoặc token không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/me/enrollments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enrollments"
                ],
                "summary": "Danh sách khóa học đã đăng ký",
                "responses": {
                    "200": {
                        "description": "Danh sách đăng ký học",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "enrollments": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/models.Enrollment"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa đăng nhập hoặc token không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.lessonRequest": {
            "type": "object",
            "required": [
                "content",
                "order_index",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "order_index": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.CourseProgress": {
            "type": "object",
            "properties": {
                "completed_count": {
                    "type": "integer"
                },
                "enrollment_id": {
                    "type": "integer"
                },
                "percent_completed": {
                    "type": "number"
                },
                "total_lessons": {
                    "type": "integer"
                }
            }
        },
        "models.Enrollment": {
            "type": "object",
            "properties": {
                "course_id": {
                    "type": "integer"
                },
                "enrolled_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.Lesson": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "course_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "order_index": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "response.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorDetail"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer Token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}